resolve issue#109

+++ src/main/java/com/softserve/repository/TeacherRepository.java
@@ -6,6 +6,6 @@ import java.util.Optional;
 
 public interface TeacherRepository extends BasicRepository<Teacher, Long> {
     List<Teacher> getDisabled();
-    Optional<Teacher> findByUserId(int userId);
+    Optional<Teacher> findByUserId(long userId);
     List<Teacher> getAllTeacherWithoutUser();
 }
